Rename files to conform with pylint conventions.
In preparation for making the code pass pylint validation
(essentially the Python Pep8 code standard), the file names
were changed to be consistent with the pylint file naming
rules.

create-test-data-set was also renamed to conform to those rules,
even though it is a Bash script, to have consistent file names
for all files, Python or not.