Added partitioning of the original file into disjoint subsets.
The new --tp option specifies what portion (a value between 0 and 1), of the
original file rows should be used for creating training sets.  The remaining
original file rows will be used for creating validation rows.

This does not work properly with multiple runs of create_sets.py to address
system open file limits, because the partitioning is not reused across runs.

The partitioning uses the Python3 random.shuffle function. This may have a
limitation in not being able to select from any of the permutations of a set of
original file row ordinals.  Some permutations (some shufflings) may not be
available. For this reason it may be useful to investigate use of the numpy
module random number functions. From the random.shuffle documentation at
https://docs.python.org/3/library/random.html

random.shuffle(x)
Shuffle the sequence x in place.

To shuffle an immutable sequence and return a new shuffled list, use sample(x,
k=len(x)) instead.

Note that even for small len(x), the total number of permutations of x can
quickly grow larger than the period of most random number generators. This
implies that most permutations of a long sequence can never be generated. For
example, a sequence of length 2080 is the largest that can fit within the
period of the Mersenne Twister random number generator.