Initial changes to support multi-pass.
When the number of training sets and validation sets to create exceeds the
system limitations on the number of open files, it is necessary to run
create_sets.py multiple times. However, even in this case the results will not
be correct, as each run will use a different set of original file rows for
creating the training sets, and a different set of original file rows for
creating the validation sets.

This branch is for changes to have one invokation of create_sets.py perform as
many passes of the original file as necessary, using as many open files as
possible on each pass.

These initial code changes still only do one pass, but detects when the open
file limit has been reached as processes as many files as were able to be open.