Initial version supporting creation of generic sets.
The --tsc/--training-set-count option is replaced by the --sc/--set-count
option, which is used for both creating training/validation sets and for
creating generic sets.

The handling of multiple passes doesn't work properly for generic sets, since
it doesn't account for opening the original data file. As many ValidationSet
objects are created, each with an open output file, until the limit is reached.
When the original data file is opened the file open limit is exceeded.

This doesn't happen with training/validation set processing because the open
that happens to fail is for the training set. The validation set for that
training set is then also discarded, with its file closed. This leaves one
remaining possible open file for the original data file.

Updated readme.txt to improve the explanation of multiple passes of the
original file (but not the problem with multiple passes for generic sets).