Bug fix: selection set available ordinals defined once.
The available row ordinals for training sets and validation sets, or generic
sets, were defined on each file pass. This is incorrect for training sets and
validation sets. In that case the available ordinals are a partition of the
original data file, with the availble ordinals in each partition selected at
random.  No validation set should have a row that is also in any training set.
This is across all training and validation sets, independently of what pass was
used to create a particular set. So the available row ordinals for the
partitioning need to be defined exactly once.

This isn't strictly necessary for generic sets, where there is no partitioning
of the original file rows.  All data rows of the original file row are
available for generic sets, and a row can appear in more than one generic set.
However, it is redenundant to define this in each pass rather than once, and
the code structure is simpler in that case.