Performance enhancement: get_random_ordinals returns set not list.
Function SelectionSet.get_random_ordinals had been changed to return a list
rather than a set. Previously set was used because it was important to choose
row ordinals for a training set that were not also row ordinals for the
corresponding validation set. It was deemed be more efficient to use sets to
determine if a candidate training row ordinal was already a row ordinal for its
validation set. When the change was made to partition the original file row
ordinals into two disjoint partitions of row ordinals, one partition for
training sets and one for validation sets, this check was no longer needed, so
using a set rather than a list was not considered necessary.

However, the efficiency of using a set rather than a list is even more
important when deciding whether a row ordinal is one that belongs to a training
or validation set, and therefore should be written to that set. The more rows
there are for each training or validation set, the longer a membership test
will take when using a list, whereas with a set the lookup is independent of
how many elements in the set. The more rows there are in the original file, the
more isntances there will be when this performance difference occurs.

On an original file with 1,000,000 data rows, creating 5,000 training sets and
5,000 validation sets, with 100 rows each, using a list took about 8 hours and
using a set took about 47 minutes, for a speedup of about 10.